LayoutTests/platform/glib/imported/w3c/web-platform-tests/css/cssom-view/offsetTop-offsetLeft-nested-offsetParents-expected.txt


FAIL Verifies that offsetTop and offsetLeft only go up one offsetParent when there are multiple nested offsetParents. assert_equals: target.offsetTop expected 10 but got 11
